+++ packages/@tinacms/cli/src/next/commands/audit-command/audit.ts
@@ -1,13 +1,98 @@
-/**
-
-*/
-
+import prompts from 'prompts'
+import { Telemetry } from '@tinacms/metrics'
+import { neutralText } from '../../../utils/theme'
 import { resolve } from '@tinacms/graphql'
 import type { Database, Collection } from '@tinacms/graphql'
-import path from 'path'
-import { logger } from '../../../../logger'
+import { logger } from '../../../logger'
 import { assertShape } from '@tinacms/graphql'
 import chalk from 'chalk'
+
+const rootPath = process.cwd()
+
+export const audit = async ({
+  database,
+  clean,
+  useDefaultValues,
+  noTelemetry,
+  verbose,
+}: {
+  database: Database
+  clean?: boolean
+  useDefaultValues?: boolean
+  noTelemetry?: boolean
+  verbose?: boolean
+}) => {
+  const telemetry = new Telemetry({ disabled: noTelemetry })
+  await telemetry.submitRecord({
+    event: {
+      name: 'tinacms:cli:audit:invoke',
+      clean: Boolean(clean),
+      useDefaults: Boolean(useDefaultValues),
+    },
+  })
+  if (clean) {
+    logger.info(
+      `You are using the \`--clean\` option. This will modify your content as if a user is submitting a form. Before running this you should have a ${chalk.bold(
+        'clean git tree'
+      )} so unwanted changes can be undone.\n\n`
+    )
+    const res = await prompts({
+      name: 'useClean',
+      type: 'confirm',
+      message: `Do you want to continue?`,
+    })
+    if (!res.useClean) {
+      logger.warn(chalk.yellowBright('⚠️ Audit not complete'))
+      process.exit(0)
+    }
+  }
+  if (useDefaultValues && !clean) {
+    logger.warn(
+      chalk.yellowBright(
+        'WARNING: using the `--useDefaultValues` without the `--clean` flag has no effect. Please re-run audit and add the `--clean` flag'
+      )
+    )
+  }
+
+  const schema = await database.getSchema()
+  const collections = schema.getCollections()
+  let error = false
+
+  for (let i = 0; i < collections.length; i++) {
+    const collection = collections[i]
+    // Not a huge fan of querying the database from outside of GraphQL
+    // but this allows us to return the list of items unhydrated, so errors
+    // in a single document don't cause the entire list query to fail
+    const docs = await database.query(
+      { collection: collection.name, first: -1, filterChain: [] },
+      (item) => ({ path: item })
+    )
+    logger.info(
+      `Checking ${neutralText(collection.name)} collection. ${
+        docs.edges.length
+      } Documents`
+    )
+
+    const returnError = await auditDocuments({
+      collection,
+      database,
+      rootPath,
+      useDefaultValues: useDefaultValues,
+      documents: docs.edges,
+      verbose: verbose,
+    })
+    error = error || returnError
+  }
+  if (error) {
+    logger.error(
+      chalk.redBright(`‼️ Audit ${chalk.bold('failed')} with errors`)
+    )
+    // } else if (warning) {
+    //   logger.warn(chalk.yellowBright('⚠️ Audit passed with warnings'))
+  } else {
+    logger.info(chalk.greenBright('✅ Audit passed'))
+  }
+}
 
 type AuditArgs = {
   collection: Collection<true>
@@ -16,25 +101,6 @@ type AuditArgs = {
   documents: { node: { path: string } }[]
   useDefaultValues: boolean
   verbose?: boolean
-}
-export const auditCollection = async (args: AuditArgs) => {
-  let warning = false
-  const { collection } = args
-  logger.info(`Checking collection ${collection.name}`)
-
-  const format = collection.format
-  args.documents.forEach((x) => {
-    const docFormat = path.extname(x.node.path)
-    if (docFormat.replace('.', '') !== format) {
-      warning = true
-      logger.warn(
-        chalk.yellowBright(
-          `WARNING: there is a file with extension \`${docFormat}\` but in your schema it is defined to be \`.${format}\`\n\n\location: ${x.node.path}`
-        )
-      )
-    }
-  })
-  return warning
 }
 
 export const auditDocuments = async (args: AuditArgs) => {
